feat: extract presentCollection tool core into the package (chunk 1d-core)
Move presentCollection's tool definition, arg/data types, and the pure
isomorphic executor (executePresentCollection — validate + echo addressing)
into @mulmoclaude/collection-plugin's `.` entry. gui-chat-protocol is now a
peer dependency (externalized in the build).

Host adapter (src/plugins/presentCollection) shrinks to re-exports:
definition.ts keeps MulmoClaude's routing META + endpoint types but sources
TOOL_DEFINITION from the package; plugin.ts / types.ts re-export. View.vue
(which mounts the host CollectionView) and Preview.vue stay host-side for now —
the CollectionView card View + browsable pages need the frontend
host-capabilities injection layer (api / nav / assetUrl / chat / confirm /
shortcuts / notifications), tracked as the remaining 1d/Phase-2 work.

Verified: yarn typecheck (server/vue/test/packages), lint, build all green;
collection + manageCollection tests pass.
